move tests directory to external module
- fixes coverage as per https://stackoverflow.com/a/60579142
