+++ .coveragerc
@@ -1,7 +1,7 @@
 [run]
 branch = True
+include = tqdm/*
 omit =
-    tqdm/tests/*
     tqdm/contrib/bells.py
     tqdm/contrib/discord.py
     tqdm/contrib/telegram.py
