+++ setup.cfg
@@ -8,6 +8,6 @@ exclude = .asv,.tox,.ipynb_checkpoints,build,dist,.git,__pycache__
 
 [tool:pytest]
 python_files = tests_*.py
-testpaths = tqdm
+testpaths = tests
 addopts = -v
 usefixtures = pretest_posttest
