+++ tests/tests_contrib.py
@@ -2,9 +2,10 @@
 Tests for `tqdm.contrib`.
 """
 import sys
+
 from tqdm.contrib import tenumerate, tzip, tmap
-from tests_tqdm import pretest_posttest  # NOQA
-from tests_tqdm import importorskip, StringIO, closing
+from .tests_tqdm import pretest_posttest  # NOQA
+from .tests_tqdm import importorskip, StringIO, closing
 
 
 def incr(x):
