+++ tests/tests_notebook.py
@@ -1,5 +1,5 @@
 from tqdm.notebook import tqdm as tqdm_notebook
-from tests_tqdm import pretest_posttest  # NOQA
+from .tests_tqdm import pretest_posttest  # NOQA
 
 
 def test_notebook_disabled_description():
