+++ tests/tests_pandas.py
@@ -1,6 +1,6 @@
 from tqdm import tqdm
-from tests_tqdm import pretest_posttest  # NOQA
-from tests_tqdm import importorskip, skip, StringIO, closing
+from .tests_tqdm import pretest_posttest  # NOQA
+from .tests_tqdm import importorskip, skip, StringIO, closing
 
 random = importorskip("numpy.random")
 rand = random.rand
