+++ tests/tests_perf.py
@@ -11,8 +11,9 @@ except ImportError:
 import sys
 
 from tqdm import tqdm, trange
-from tests_tqdm import pretest_posttest  # NOQA
-from tests_tqdm import importorskip, skip, StringIO, closing, _range, patch_lock
+from .tests_tqdm import pretest_posttest  # NOQA
+from .tests_tqdm import importorskip, skip, StringIO, closing, _range, \
+    patch_lock
 
 
 def cpu_sleep(t):
