+++ tests/tests_tqdm.py
