+++ tox.ini
@@ -14,9 +14,8 @@ deps =
     coverage
     coveralls
 commands =
-    pytest -v --cov=tqdm -k "not tests_perf" tqdm/
+    pytest -v --cov=tqdm --cov-report=xml --cov-report=term -k "not tests_perf"
     - coveralls
-    - coverage xml
     - curl -OL https://coverage.codacy.com/get.sh
     - bash get.sh report -r coverage.xml
 allowlist_externals =
@@ -28,10 +27,9 @@ deps =
     {[coverage]deps}
     codecov
 commands =
-    pytest -v --durations=10 --cov=tqdm -k "not tests_perf" tqdm/
+    pytest -v --durations=10 --cov=tqdm --cov-report=xml --cov-report=term -k "not tests_perf"
     - coveralls
     codecov
-    - coverage xml
     - curl -OL https://coverage.codacy.com/get.sh
     - bash get.sh report -r coverage.xml
 allowlist_externals = {[coverage]allowlist_externals}
@@ -81,13 +79,13 @@ deps =
     {[extra]deps}
     tensorflow
 commands =
-    pytest -v --durations=10 tqdm/tests/tests_keras.py
+    pytest -v --durations=10 -k tests_keras
 
 [testenv:perf]
 deps =
     pytest
 commands =
-    pytest -v --durations=10 tqdm/tests/tests_perf.py
+    pytest -v --durations=0 -k tests_perf
 
 [testenv:flake8]
 deps = flake8
